Move _filter_config to new config submodule
